[Support] Add walkPostOrder and walkInversePostOrder to InstanceGraph (#8974)
Add new helper functions to the `InstanceGraph` that implement a full
post-order and inverse-post-order walk across _all_ modules. This pulls
a common pattern into the instance graph itself, where a user would
iterate over the graph's nodes and then use an `llvm::post_order_ext`
iterator to not revisit prior nodes with an externally-held set.

The instance graph implements the graph traits, which makes using
`llvm::post_order` it tempting. But this is usually not what we want.
The graph does not have a unique root node. Instead, the FIRRTL flavor
of the graph uses the "top" module as specified by the circuit, and the
HW flavor of the graph uses a dummy node that points to all _public_
modules. In most passes we want to iterate over _all_ modules in the IR
with a specific order, not just a subset of the modules. These new
helper functions make this common use case more obvious and may allow us
to remove the graph traits from `InstanceGraph` again in the future.

Most of this PR is just mechanical identation changes. One brittle test
for instance choice needs to be updated since that relied on the fact
that the dedup pass would ignore uninstantiated modules.