[ImportVerilog] Fix single argument expressions in severity tasks
This commit fixes missing support for expression evaluation in severity system tasks
as well as is the `$display` and `$write` tasks.

According to IEEE 1800-2023 Section 21.2.1 "The display and write tasks", it is legal
to pass fully formatted expressions as the arguments of severity, display, and write tasks.
In the previous implementation, it was assumed that the first argument would always be
a StringLiteral in need of formatting; this commit adds a case distinction in the handling
of these statements that ensures we first check whether we need to treat the arguments
as a format string instance, or we are being handed a fully formatted string.