[FIRRTL] Ensure all types and attributes are walkable
Upstream MLIR requires type and attribute storages to define a
`getAsKey` function, and for the return type to be something that
implements the corresponding attribute/type walk. Most of our FIRRTL
types and attributes predate these requirements. This PR is mostly busy
work, adding `getAsKey` to the types and attributes that don't yet have
it, and changing `std::pair` to `std::tuple`. (For some reason upstream
type/attribute walking works on tuples, but not pairs.) Also add unit
tests for attributes and types that ensure the walks actually work.

These tweaks allow circt-reduce to be a lot more clever about which ops
it touches, because it can see a lot more symbol uses now. I also expect
the `InnerSymbolDCE` pass to potentially benefit from this, at least if
inner symbols are weirdly nested within FIRRTL attributes and types.

Most of this was done by Claude Sonnet 4, with a lot of handholding.