[ESI] Route simulator compilation/exec output through callbacks
To better support "library mode"-usage of these scripts, we need a way for users to get a hold of the stdout/stderr from the compilation and simulation parts of the Simulator class.

This PR changes the compile/sim logging to be performed through a callback, as opposed to directly to a file descriptor. If no callbacks are provided by the user during Simulator construction, the default (file-based) logging will be used.