[ESI] Allow single-file additions to SourceFiles
Extend `SourceFiles` to also allow single file additions to the source file list.

Small, but important addition, due to the perniciousness of verilog compilers(/simulators), in that e.g. type-files have to have a strict ordering when they have inter-dependencies. Just scanning a directory and adding them in lexical order may or may not (...obviously) cause problems in that aspect. With this change, gives users control over the order in which these files are added to the source list.