fix: prevent canonicalize looping for extremely long time
A hanging bug can manifest with very large inputs to minimize at
canonicalize pass. This is because if the maxNumRewrites is not set,
it defaults to kNoLimit, which could be an arbitrary large number.
This optional command line argument makes it possible to set the max
number of rewrites for large input files to minimize.

Signed-off-by: Tianrui Wei <tianrui@tianruiwei.com>