[FIRRTL] Gate class deduplication behind an option
Add a `dedupClasses` option to the Dedup pass, enabled by default. When
the option is enabled, classes are deduplicated like modules, which
makes their names unstable. Some current flows that use firtool require
this behavior. If the option is disabled, classes are considered
nominally typed (which they are intended to be), and are not
deduplicated at all.

Add a corresponding `--dedup-classes` option to firtool, also enabled by
default. This allows flows to continue relying on the current class
deduplication behavior, but they can opt into proper nominally typed
classes as they see fit.

Eventually we'll want to remove these options and make classes *not*
dedup unconditionally. @dtzSiFive calls this a compiler chicken bit.