[Support] Guard ScopedDebugPassLogger in release
Fix a nightly regression failure [[1]] where the `isCurrentDebugType`
function has a different signature when NDEBUG is/is not defined.  (I
think this is actually an upstream issue.)  Avoid this entirely by
guarding the definition of `ScopedDebugPassLogger` which then forces users
to use the macro `CIRCT_DEBUG_SCOPED_PASS_LOGGER` which is better anyway.

[1]: https://github.com/llvm/circt/actions/runs/18192317432/job/51790002617

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>