[LLHD] Add new RefType to replace hw::InOutType
Add a `RefType` to the LLHD dialect and replace all uses of HW's
`InOutType` across the LLHD dialect and the ImportVerilog pipeline. We
have been abusing the inout type to mean "signal/variable reference",
which becomes problematic if we are trying to place `!llhd.time` or
`!llvm.ptr` values behind such a reference. (The HW inout type requires
the inner type to be a plain hardware value.)

This change now allows the circt-verilog frontend to generate variables
with the `time` type and treat them just like any other type.

Most of the changes here are just systematic name replacements. I have
also removed a few `qualified(...)` wrappers around types and made sure
that when an op is known to accept or return a ref type, the printer
would just print `<T>` instead of `!llhd.ref<T>`. This is the default
behavior for unambiguous types, but sometimes this gets lost when
explicitly askign for qualified types, or using `functional-type`.