[FIRRTL] Add early exits to LowerDomains
Add early exits to the LowerDomains pass to avoid recreating modules and
instances or updating their attributes if there are no domains to lower.

Note: eventually this should be replaced with checks that all domain
information is known.  I.e., that InferDomains has run.  Until this
happens, this kind of early exiting (which is the behavior of the pass
today, just with potentially cloning ops and types) is beneficial from the
perspective of inserting the pass unconditionally into the firtool pass
pipeline.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>