[FIRRTL] Add LowerDomains to firtool pipeline (#9099)
Add the LowerDomains to the low FIRRTL to HW pipeline inside firtool.
This is the latest position in which it can run as it has a hard
dependency on needing to run before LowerClasses.

This is not conditionalized with a chicken switch option as this pass will
eventually be required for correctness and it is a no-op if no domain
information exists in the circuit.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>