[FIRRTL] Make containingModule compatible with Chisel changes (#9126)
Adjust the LowerClasses pass to make it match the `containingModule`
field more generously. Soem Chisel-side tweaks are going to change this
name from `containingModule_in` to `__in_containingModule`, which the
current code does not match.

It's a bit unfortunate that we have this weakly stringly-type interface
in the compiler, but we need it at the moment for compatibility.