[Moore][ImportVerilog] Add support for class method forward decls
This commit adds the plumbing to convert `slang::ast::MethodPrototypeSymbol`s corresponding to class
method forward declarations. This is done by looking up the implementation and converting it when
encountering a forward declaration. This makes sure field order within the class declaration is
respected and the emitted `methoddecl`'s signature is correct.