[ImportVerilog] Fix capture callback chaining
This commit fixes the chaining sequence of capture callbacks.
Previously, parent callbacks would be run before the callbacks of the inner scope;
this could lead to the callback of the inner scope rejecting a capture but the outer scope
accepting it. This commit first checks whether the inner scope accepts the capture
before also running the outer scope's callback.

This commit further rejects non-reference captures (these should not happen on reads/writes anyway).