[HWToLLVM] Do not take illegal shortcut in `array_get` lowering. (#9172)
Currently, the LLVM lowering of `hw.array_get` will attempt to reuse pointers to array values to avoid rematerializing the array on the stack. This is only legal as long as we can ensure the underlying buffer remains unchanged. It works within the confines of the HWToLLVM lowering, which always allocates a new buffer when modifying an array. However, operations of other dialects (e.g., `arc.state_write`) might still write to the pointed-to buffer between the `load` and the `array_get`, causing it to return an incorrect value.