[ImportVerilog] Add global variable support
Add new `moore.global_variable` and `moore.get_global_variable` ops to
the Moore dialect, which define global variables as a symbol and return
a reference to a global variable given its symbol, respectively.

Add support for variables at the root level and within packages in SV
inputs. Both of these get converted to `moore.global_variable`s. When
referring to a global variable by name, create `get_global_variable` ops
to resolve the symbol name to a `!moore.ref<T>` which can then be read.

Lowering to the core dialects is not yet supported. We very likely need
an equivalent global signal construct in LLHD.