[tools] Add domaintool
Add a new command line program, domaintool, which is intended to be used
to consume domain information in a final MLIR blob and produce constraint
files, e.g., SDC files.  Currently, this is targeted at generating a
SiFive-internal format ("Clock Spec JSON") which we then convert to SDC
files.  This is an intermediate step while this is stood up and will
eventually be removed.

This includes a basic, in-tree registration mechanism to register new
handlers.  This is intended to allow for us to extend this to add actual
SDC and Unified Power Format (UPF) handlers.

Briefly, this tool works by providing domains and their parameters on the
command line as well as assignments of the provided domains to specific
ports.  An example invocation is the following:

``` shell
domaintool \
  --module Foo \
  --domain ClockDomain,ClockA,10 \
  --domain ClockDomain,ClockB,20 \
  --assign 0 \
  --assign 1 \
  Foo.final.mlir
```

This instantiates two clock domains, with different names and periods.  It
then hooks up the first domain, ClockA, to the first domain port and the
second domain, ClockB, to the second domain port.  When run, this then
generates JSON output due to the ClockSpecJSON handler being registered:

``` json
{
  "clocks": [
    {
      "name_pattern": "ClockA",
      "define_period": 10,
      "clock_relationships": []
    },
    {
      "name_pattern": "ClockB",
      "define_period": 20,
      "clock_relationships": []
    }
  ],
  "static_ports": [],
  "asynchronous_ports": [],
  "synchronous_ports": [
    {
      "name_pattern": "ClockA",
      "name_pattern": [
        "a"
      ],
      "comment": null
    },
    {
      "name_pattern": "ClockB",
      "name_pattern": [
        "b"
      ],
      "comment": null
    }
  ]
}
```

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>