[domaintool] Add `-assign u` support
To work around the problem of needing to evaluate a FIRRTL "ABI" class
that contains domain-lowered properties _and other non-domain properties_,
add the ability to instantiate an object with an "unknown" value
assignment via `-assign u`.

Previously, `domaintool` was unusable if the class being instantiated
contained any non-domain information.  This occurs in any internal,
non-trivial design.  This provides a mechanism, using the "unknown" value
added in #9247, to work around this.

Note: as mentioned on #9247, it would really be better if there was a way
to define a "domain function" that contained only the information that we
cared about.  However, the FIRRTL "ABI" for properties was not implemented
this way and only a single entry point is provided.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>