[ci] Change static build to use Rocky Linux 8
Switch the static build from "true static" via muslc/Alpine to use "mostly
static, but using an old 2.28 glibc" by building in a Rocky Linux 8
container.

This replaces the fully static Alpine Linux + muslc (and plethora of
workarounds to make this work) due to observed issues with Arcilator and
LLVM-JIT.  While this was nice if we could make this work, I have come to
the conclusion that it was too much trouble to make this work.

Note: if we do ever need an even older glibc ABI, we _can_ build on CentOS
7.  This was the original plan with this commit.  However, CentOS 7 is
end-of-life and it's more complicated to make it work, e.g., the standard
package repositories are not available.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>