[MooreToCore] Fix lowering of dynamic extracts of nested arrays (#9257)
Fixes the lowering of single element dynamic extracts from arrays of arrays in MooreToCore. Previously, an array typed result of an extract would always cause the emission of a slice operation instead of a get operation. This changes the logic to emit a get operation if the result type is equal to the input array's element type.