Revert "[ci] Change static build to use Rocky Linux 8"
This reverts commit 7982bdf4df838efb5d9e35cbf9b6ff7b2e725028.

Thanks to the fix by @fzi-hielscher in 0e1abe7fe, the muslc builds will
continue to work, by not running Arcilator JIT tests.  We think that there
is line of sight to enabling these, though.

In the future, we may revisit switching to Rocky Linux for publishing.
However, we don't need to do this right now.