[FIRRTL] Remove -allow-adding-ports-on-public-modules.
This controls whether to allow "wiring problems" to bore
through public modules, adding new ports to the interface.

This was a transition option, and the statistic was used
to track how many were left to fix.

Most wiring is now performed in Chisel.
Remaining wiring mechanisms:
* memory taps
* tapping for GC views
    Views already bore their own signals in Chisel,
    but GC Views code is still capable of wiring.
* legacy wiring

This option hasn't been used in flows for over a year.