[circt-synth] Disable test if yosys-abc unavailable
I ran into a situation on my machine where yosys was installed, but
yosys-abc was not. Since the `circt-synth/abc.mlir` test only requires
yosys-abc, add a separate feature flag and perform and additional check
to see if yosys-abc is installed alongside yosys. If it is not, disable
that test.