[OM] Remove Pure from ObjectOp to preserve identity
Remove the `Pure` trait from `om.object`. With the op marked pure, the
CSE pass will combine instances of the same class into a single
instance. This is problematic, since users may expect distinct instances
of the same class to be preserved as two separate instances instead of
being deduplicated.

Instead, add the `MemAlloc` trait to the result of the `ObjectOp`. This
causes the op to no longer CSE, but it allows unused objects to be
erased.