[FIRRTL] Allow domain ports after associations
Change FIRRTL parsing/printing (both text and Dialect) to allow for
domain ports to show up after ports which are associated with them.  E.g.,
make the following legal and round-trip correctly:

    module Foo:
      input a: UInt<1> domains [A]
      input A: Domain of ClockDomain

This is done as if this is not allowed, then the creation of domains by
the `InferDomains` pass will need to do port reordering.  Additionally, we
can't think of a good reason to actually disallow this.

This was already allowed (not rejected by a verifier) in FIRRTL Dialect.
However, it just could not be parsed/printed.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>