[Verif] Add hw.constant Verif canonicalizers
This is a follow-on to e4c0d615f that adds additional canonicalizers for
Verif assert and assume operations where the predicate is a `hw.constant`
true.  This is in addition to the existing canonicalizers that work with
the new constant property.

This should fix some observed missed optimizations that caused
`assert(1'h1)` to show up in some internal designs.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>