[integration_test] XFAIL formal circt-tests
Mark two circt-tests as XFAIL as these have issues with the new,
optimizations around assertions.  Specifically, assertions can now be
optimized away if they are known to the compiler to be true.  However,
tools like `sby` seem to not accept empty circuits.  An attempted
workaround of using `cover(true)` instead of `assert(true)` seems to work
for `sby`, but appears to not be supported by `circt-bmc`. See discussion
about this [[1]].

[1]: https://github.com/llvm/circt/pull/9325#issuecomment-3638322059

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>