[Verif] Add hw.constant Verif canonicalizers (#9325)
This is a follow-on to e4c0d615f that adds additional canonicalizers for
Verif assert and assume operations where the predicate is a `hw.constant`
true.  This is in addition to the existing canonicalizers that work with
the new constant property.

This should fix some observed missed optimizations that caused
`assert(1'h1)` to show up in some internal designs.

Mark two circt-tests as XFAIL as these have issues with the new,
optimizations around assertions.  Specifically, assertions can now be
optimized away if they are known to the compiler to be true.  However,
tools like `sby` seem to not accept empty circuits.  An attempted
workaround of using `cover(true)` instead of `assert(true)` seems to work
for `sby`, but appears to not be supported by `circt-bmc`. See discussion
about this [[1]].

[1]: https://github.com/llvm/circt/pull/9325#issuecomment-3638322059

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>