[FIRRTL] Don't create empty modules in LowerLayers
Improve the LowerLayers pass to not create empty modules lowered from
empty bind layer blocks.  Because of the way that bind layer blocks are
lowered to modules, which must not include instantiations (to avoid
Verilog's illegal bind-under-bind), whenever a bind layer block is known
to be empty, we can just skip it.

Note: this does not remove the layer ABI file that is created, as this is
mandatory in the ABI.  However, it makes the layer ABI file a no-op.

Fixes #9331.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>