[ImportVerilog] Fix class property segfault and emit error
Fix an issue around class property handling where code would not
properly catch failures in the `getAncestorClassWithProperty()` and some
other functions, and would instead try to access the null result.

Also add a missing error message if a class property cannot be resolved.

This caused about half of the segfaults in the sv-tests suite [1].

[1]: https://github.com/circt/circt-tests/blob/results/2025/2025-12-15-131140-main-82227de/sv-tests/segfaults.txt