[ImportVerilog] Add support for program definitions (#9377)
Programs are essentially just modules. There are a few subtle semantic
differences, but we aren't far enough down the simulation path to have
those show up. For the time being, just treat programs as modules. We
will add the necessary exit blocking mechanisms later on when we have
sufficient simulation support in Arcilator.